Rename Location to SpacetimeControl.
