+++ app/debugger/modules/variableHijacker/components/configs/theme.ts
@@ -41,7 +41,7 @@ const icon = {
   WebGLRenderer: { img: `${iconPath}/class.png`, backgroundColor: pink },
   // General internal classes
   ChangeTracker: { img: `${iconPath}/ChangeTracker.png`, backgroundColor: darkAqua },
-  '[Internal]Location': { img: `${iconPath}/class.png`, backgroundColor: darkAqua },
+  '[Internal]SpacetimeControl': { img: `${iconPath}/class.png`, backgroundColor: darkAqua },
   // Icons not directly related to classes
   sphere: { img: `${iconPath}/sphere.png`, backgroundColor: defaultColor },
 };
